requirements.txt
libusb1==3.1.0
mypy==1.11.2
mypy-extensions==1.0.0
ndeflib==0.3.3
nfcpy==1.0.4
pyDes==2.0.1
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pytz==2024.1
shiboken6==6.7.2
SQLAlchemy==2.0.34
tomli==2.0.1
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
